Updated PR comments
