+++ Server/AdditionApi/DB/AppDbCOntext.cs
@@ -1,0 +1,7 @@
+using Microsoft.EntityFrameworkCore;
+
+public class AppDbContext : DbContext
+{
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
+    public DbSet<Record> Records  => Set<Record>();
+}
