+++ Server/AdditionApi/DB/Record.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+[Table("Record")]
+public class Record
+{
+    [Key]
+    public string Key { get; set; } = string.Empty;
+    public string Value { get; set; } = string.Empty;
+}
